Revert "Merge branch 'TestScene' into feature/teilenDerBlickrichtungHololensAndVive"
This reverts commit 597a355d7d9f5ee1ed05ee3f876ce9b8be2feff8, reversing
changes made to 7ab36cb737d2cca789266c930fcc663f32bddab2.
