+++ Frontend/VIAProMa/Assets/Prefabs/Multiplayer/Resources/Avatar Networked.prefab
@@ -481,15 +481,15 @@ PrefabInstance:
       propertyPath: m_RootOrder
       value: 10
       objectReference: {fileID: 0}
-    - target: {fileID: 7619380334928439537, guid: fddad501a37828346aa422a968ca8973,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 12
-      objectReference: {fileID: 0}
     - target: {fileID: 7619380334928439539, guid: fddad501a37828346aa422a968ca8973,
         type: 3}
       propertyPath: m_RootOrder
       value: 11
+      objectReference: {fileID: 0}
+    - target: {fileID: 7619380334928439537, guid: fddad501a37828346aa422a968ca8973,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 12
       objectReference: {fileID: 0}
     - target: {fileID: 3310419911003119975, guid: fddad501a37828346aa422a968ca8973,
         type: 3}
