--- Frontend/VIAProMa/Assets/Tests/GazeShareTest/GazeShareOnClick.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62ef73b291070f84ebd97a7ee8806453
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
