--- Frontend/VIAProMa/Assets/Tests/GazeShareTest/leftTestCube.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8c35bb72175c1fc479235cd8783b081d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
