--- Frontend/VIAProMa/Assets/Tests/GazeShareTest/rightTestCube.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 104e0ab10e8291242a27895796b39648
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
